Add explicit libkf5plasma5 versioned dependency
This needs to be handled by the cmake parse script. A raw draft would be to convert the
cmake requirement in a debian source package, with dev dependencies and runtime dependencies.

That would also need to handle soversion bumps by removing the "old" runtime dependencies.

Closes: 829614
Thanks: Gard Spreemann and others for reporting and the updates